Implement new signal handling scheme to protect IMDB from signals.
For all IMDB using processes:
   sipXproxy
   sipxacd
   watchdog
   mediaservercgi
   replicationcgi
   sipxrls
   sipstatus
   sipxregistrar
   sipxpark

Block the main thread (and all subsequent threads) from catching 
signals.  Create a signal handling thread that is the only one 
that has signals unblocked.  Wait on sigwait() in that thread for 
external signals.  The signal thread then sets the global shutdown 
flag that the main thread monitors.

This is the best practice for handling signals in a POSIX threads 
environment.  No other thread will ever be interrupted due to a signal.  
This prevents the IMDB (and any other thread) from ever being interrupted 
while in a critical section.  It also prevents any system call from
being interrupted and returning EINTR.

Things to note:
  SIGPIPE is set to be ignored in various deep libraries. This
is correct behavior.  SIGCHLD was being caught by a reaper
handler in OsProcessLinux.cpp.  I changed this to just ignore
SIGHLD, which as of POSIX-1.2001 will automatically reap children.

  INTERNAL signals (like abort(), SIGSEGV, SIGBUS, SIGFPE) will 
still occur, and cause their default handling to occur--namely
core dumping.  This could occur during an IMDB critical section.
But catching them and trying to clean up the IMDB (as the code
used to do) is prone to deadlock anyway.

Other major changes:

   OsProcessLinux::launch() was changed to close all file
descriptors except for STDIN, STDOUT, and STDERR.  If this isn't done
the children inherit open files from the parent process.  In
watchdog's case, this included the IMDB and a control socket.  In 
addition, the child's signal mask is unblocked just before the
exec().  This prevents the child from inheriting a blocked signal
mask from the parent.

   sipxrls has an initial long wait before starting.  It assumed
that wait would be interrupted by the shutdown signal.  While
this is a bad assumption in general, it is sure to fail with this
new scheme.  I replaced the long wait with a loop of waits that
also checks the global shutdown flag.

   watchdog.  As of 3.9 watchdog is started as root, and it
does a setuid() to drop privledges to SIPXUSER.  This was done
after the logging was started, so that it turned out the watchdog.log
was owned by root.  Once the privledges were dropped, when the
logging subsystem tried to re-open the log it could no longer 
write to the log, so all subsequent logging failed.  I moved the
dropping of root privledges to before the log is created.  In addition,
I added a setgid() to SIPXGROUP so all the sipX processes are in
the correct group.  This fixed the need to gdb processes as root.

   mediaservercgi and replicationcgi are now immune from all external
signals.  Previously they would "close the IMDB" but not exit on
catching a signal.  Now it just ignores all signals.  As the CGI
scripts aren't supposed to block, and there is no way to signal them
anyway (as they are controlled by httpd), this seemed the safest
course.

   statusserver used to only the monitor the global shutdown flag
every 30 seconds.  All the other processes use 1 or 2 second delays.
I changed it to a 2 second delay to match.

   OsTaskWnt.cpp was changed to match the OsTaskLinux.cpp call 
signatures, but the code does nothing.  Windows ports and other
non-POSIX ports will need to do the right thing for their platform.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/3.10@12123 ab1d8caa-1f67-47f1-9e81-24633a41865c