XCF-2397 only disable MRTG when last copy of the sipxconfig-mrtg is erased
When a new version of RPM package is installed the older version does not get removed until the newer version is fully installed.
The sequence is:
%pre for new package
%post for new package
%preun for old package
%postun for old package

since sipxconfig-mrtg was disabling monitoring in preun section, updating it was removing MRTG from sipXconfig UI.
Added checking for the number of packages installed to prevent it.

Unfortunately it means that upgrading from previous (broken) version of sipxconfig-mrtg will still disable monitoring.
Upgrading from this version will work OK.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/3.10@12140 ab1d8caa-1f67-47f1-9e81-24633a41865c