Increase the maximum number of per-process file descriptors
For larger systems, default limit may not be enough.  This
change tries to increase the per-process file desciptor maximum
to 16384 descriptors per process.  If it fails, it's only a warning.

This is done in the /etc/init.d/sipxpbx startup script which is
run as root.  All processes started by that script should inherit
the larger value.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/3.10@12150 ab1d8caa-1f67-47f1-9e81-24633a41865c