Change awaitSignal() to internally handle EINTR
An bug in older glibc implementations allows sigwait() to
return EINTR incorrectly.  Handle this case by restarting
the sigwait if EINTR is returned.

glibc change:
http://sourceware.org/ml/glibc-cvs/2005-q3/msg00382.html

This change was picked up in RH4 Update 2.

I also reverted the previous change the awaitSignal() to only
check for "standard" signals, as that turned out not to be
the root cause.

Fixes XECS-1296 - watchdog error when restarting services after 3.10.1 upgrade

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/3.10@12215 ab1d8caa-1f67-47f1-9e81-24633a41865c