Call _exit(2) instead of exit(3) on media server and ACD processes
Fixes XECS-1379 - threads that are accessing static objects cause core dumps on exit() when the objects are destructed, but the threads are still running.

I tried really hard to get all the threads in the media server and ACD
to stop before calling exit(), but too often those threads wouldn't
stop cleanly.  See comments in sipXvxml/src/client/ivr/clientMain.cpp

In the end I gave up and decided to just call _exit(2) to avoid the 
atexit() processing which will prevent the destruction of the static objects.
The threads will die when the process exits.

Probably the other media processes like sipxpark and sipxpresence should
get this fix as well, but they don't seem to have a problem with shutdown
like the media server and ACD do.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/3.10@12503 ab1d8caa-1f67-47f1-9e81-24633a41865c