XECS-1376 change the name of callresolver log file (to sipxcallresolver.log)
callresolver is using standard Ruby Logger utility for logging. Logger does not reopen log files
periodically. In such cases logrotate needs to be told that it should 'copytruncate' the log files.

sipx configuration for logrotate will 'copytruncate' log files which names start with 'sipxc'. After
the name change sipxcallresolver.log matches that pattern, and logging works correctly. (And
incidentally it is a better name for callresolver log anyway.)

Better fix would be to handle signals that tell application to close and re-open its log files. That
needs to be done for sipXconfig as well (possible other applications that use log4j for logging).


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/3.10@12546 ab1d8caa-1f67-47f1-9e81-24633a41865c