XCF-2476 fix for restarting ConfigServer from Services page
In order to help with IMDB locking problem sipXconfig changed the way it restart services (replacing
'restart' with 'stop' followed by 'start'). The new restarting strategy does not work for sipXconfig
itself, since once it is stopped there is nothing in the system that would start it.

Added a special case for sipXconfig in manageServices restart handling: it's using RESTART command.
Since sipXconfig does not bind to IMDB it should not interact with the previous fix (XECS-1257).

(Kevin) - Modified restart code to restart config server last, based on Misha's comments in the Jira issue


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/3.10@12645 ab1d8caa-1f67-47f1-9e81-24633a41865c