XCF-2386 add pager to user phonebook
Add a pager component to the existing table. Use tableVariables.tablePageSize to obtain the default
size of the table.
Make query parameter persistent so that search stick when user changes pages.
Initialize tableModel lazily: in some cases pager will request the model before pageBeginRender is
called, we work around it in a similar way on other pages.

3.10.2 specific (slightly more conservative fix than in mainline where I used the newly updated
Table component)


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/3.10@12667 ab1d8caa-1f67-47f1-9e81-24633a41865c