XECS-1594 - Stranded MOH calls
Implement keepalive mechanism.  Use re-INVITEs for the park server and OPTIONs for the MOH server.
Any non-2XX response to the re-INVITE and a 408, 404, or 481 response to the OPTIONs will disconnect the
call. The keepalive messages are sent every 300secs by default and can be
adjusted by passing the parameter "SIP_PARK_KEEPALIVE_TIME" and its value in
secs to the sipXpark configuration file.

Patch from Arjun Nair - thanks.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/3.10@13975 ab1d8caa-1f67-47f1-9e81-24633a41865c