XCF-2777 fixed generating polycom directory files
If the same contact/number field is present in phonebook directory and in speed dial list, speed
dial takes a precedence (phonebook entry is not generated).
It prevents phone book entries (controlled by administrator) from eclipsing speed dial entries
controlled by the user.
This is a workaround for a Polycom problem: the phone should allow for multiple entries with the
same contact in a directory list.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/3.10@13995 ab1d8caa-1f67-47f1-9e81-24633a41865c