XECS-1207: Backup fails on SuSE due to permissions
Checking for a postgresql process now attempts to locate a valid pid file,
and if it exists then verifies the presence of the pid contained within 
against the active process list. If there is no valid pid file or if the
pid file is not readable by the user attempting to run the backup, then
the check for a postgresql process reverts to checking the active process
table for a running postgresql instance (process name = postmaster).

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@14231 ab1d8caa-1f67-47f1-9e81-24633a41865c