XCF-2825 fix adding and selecting SBCs in Internet Calling screen
New method in SvcDeviceManager which checks if the max limit is reached

The SbcDeviceSelect component uses this method when creating the model for the drop-down list and if
the limit is not reached the possibility to add new sbc of the corresponding type is added in the
combo.

Added also some code which allows designating sbc type as not visible in some certain pages
('internetCallingSupported'). For example, the internal sbcs are not visible on the Internet Calling
page nor to add (the existing ones) nor to create new ones.

Patch from Bogdan Brezoi - thanks!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@14257 ab1d8caa-1f67-47f1-9e81-24633a41865c