XCF-3219 Add option to Include Phone/Gateway profiles on Snapshot screen (sipx-snapshot)
Phase 1 of the implementation is the UI layout change, which adds headers for the existing items.
The main reason for doing this is to make it apparent that the Configuration "Credentials" does not
apply to Phone and Gateway Profiles.  This change does not actually include any Phone and Gateway
Profiles content, and therefore does not depend on XECS-2056: Include Phone/Gateway Profiles in
sipx-snapshot.

Migrate SnapshotPage: .page -> .html

Patch from Paul Mossman - thanks!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@14526 ab1d8caa-1f67-47f1-9e81-24633a41865c