XCF-3240: Fixed the auto-refresh interval so that changes are immediately reflected.
Tapestry doesn't have great support for triggering client-side JavaScript when a
component is updated, but it does provide a way to send a "status message" to a Dojo
"topic". By attaching a listener function to the topic on the client side, the component
can update the JavaScript by sending the new interval value as a status message on the
"refreshInterval" topic. This allows the refresh timer to be reset.

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@14562 ab1d8caa-1f67-47f1-9e81-24633a41865c