XCF-2690: Added the AutoRefreshForm to all the conferences pages.
To accomplish this, a few tweaks had to be made to the AutoRefreshForm.
Some pages that show conferences have multiple tabs, and already are
contained by a form. That form is changed to be an AutoRefreshForm, but
this results in having a "Refresh" button and an auto refresh checkbox
on every tab, even where this is not applicable.

The changes I've made to AutoRefreshForm allow the "refresh" UI to be
omitted (via the 'showRefresh' parameter).

Additionally, changing to the async AutoRefreshForm broke deleting
conferences - the 'selected' and 'tag' properties were not being updated
properly, so I switched the button for deleting conferences to use a
standard listener method.

Finally, a bit of refactoring had to be done for all the UI tests to
pass.

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@14618 ab1d8caa-1f67-47f1-9e81-24633a41865c