Correct ACD reports SQL
The Agent Activity Summary Report SQL statement was using a join, 
which created a cross-product of the two tables, giving wildly incorrect 
results.  Changed this to a subselect instead of a join.

Also changed the time interval selection criteria: old report was
Checking for "time > timestart", I changed it to "time >= timestart"
so that the start time is included, and the stop time is excluded.  This
prevents dropping data that matches timestart exactly.

Fixes XCF-2733 - ACD Historic Reports: Inconsistent numbers on Agent Activity Summary report

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@14662 ab1d8caa-1f67-47f1-9e81-24633a41865c