XCF-3405 replicate datasets to remote location
IMDB datasets need to be replicated before sipXconfig attempts to start
any services on remote location. They used to be replicated in
ReplicationTrigger: however that was not reliable (no guarantee that
replication is completed before the services are started). Now
ServiceConfigurator provides initLocations method that can be used to
push all datasets and files that are not part of the standard service
replication.

Additionally changed the defaults for
ReplicationTrigger.replicatieOnStartup. sipXconfig will not replicate
datasets unless 'first-run' flag is set, or 'replicateOnStartup' is
explicitly set to 'true'.

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@14840 ab1d8caa-1f67-47f1-9e81-24633a41865c