sipXpage: Try not to page the originator of the page
Accepted rjoly's patch for XECS-2232 - when paging from any phone in the group being paged, it places the page on hold and has to be retrieved before page can be completed. _paging works fine when paging any other group.

The actual reported problem turnd out to be a misconfigured phone, but that 
doesn't mean we cannot at least try to skip paging the originator.  This 
patch uses the From: header and if the user@host happens to match one of the 
users in the page group, it will skip paging that user.  There are lots of 
reasons why this might not match when it "should", but it's about as good 
as we can get from the vagaries that is SIP.

I added some additional logging and general cleanup as well, including
Eclipse project settings.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@14855 ab1d8caa-1f67-47f1-9e81-24633a41865c