sipXvxml: Fix race condition on hangup
Whenever OpenVXI went to start looking for DTMF or start a recording,
if the call had already detected a hangup condition it would return
VXIrec_RESULT_DISCONNECT.

But since forever, the next layer up expected VXIrec_RESULT_SUCCESS
or it would end the call due to an error.  Most of the time, this
isn't even a problem, as hangup should end the call.  But, if there
was a recording pending, then bad things(TM) would happen.

If the hangup occured during the recording, normally it would do
the "right thing".  However, if the recording has just stopped by 
itself (due to the silence timeout or even a digit press), and then
the hangup occured in a certain timing window, when it went
to collect the next set of digits, the collection could fail with
the VXIrec_RESULT_DISCONNECT return value, which would cause 
OpenVXI to skip the hangup processing of saving the file, and just 
abruptly end the call.

This change looks for VXIrec_RESULT_DISCONNECT as a return value,
and then returns the GrammarManager::Disconnect value, instead of
GrammarManager::InternalError.  The former causes hangup processing
to occur, the latter the call to be dropped without prejudice.

Why this should suddenly show up now, I don't know.  It's been in the
code for at least the last 5 years (as far back as I could find).
Perhaps something is now faster or slower than before, and is 
exposing the window.  As it was, I had a hard time getting it to happen
consisently before this change.

Fixes XECS-2320 Voicemail mesage "disappears" on hang-up.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@14906 ab1d8caa-1f67-47f1-9e81-24633a41865c