XECS-2172: Need ability to get logs past 1 day old from Snapshot Screen
Modified sipx-snapshot to cycle back through rotated log files searching
for log entries in the time span. Modified sipx-log-cut to handle the
case where the time span was not entirely contained in the log file, and
to return an exit value of 2 when the search end time is earlier than
the last log entry in the input file. Also added unit test cases for
sipx-log-cut.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@14945 ab1d8caa-1f67-47f1-9e81-24633a41865c