XCF-3614 fix errors during DB upgrade phase
There is an issue with properly saving the location when we publish
location save event during DB init phase.
The underlying problem is interesting: we somehow end up with 2 detached
objects associated with the same ID and get infamous
NonUniqueObjectException from Hibernate.
It would be great if we can find a root cause and fix it but we actually
should not publish 'onSave' and 'onDelete' events during DB init phase
at all.
So instead of tracking down the root cause I am just making it sure here
we stop publishing unnecessary events.
- change name of LocationManager mathod from 'save' to 'store' since all
  save methods publish events
- use storeMigratedLocation in LocationSpecificServiceInitTrigger: we
  need save events there either

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@15152 ab1d8caa-1f67-47f1-9e81-24633a41865c