FreeSWITCH: Be more aggressive shutting down FreeSWITCH
If you try to stop FreeSWITCH before it is finished starting, 
it doesn't stop.  It takes about 30 seconds for FS to be started 
enough accept the stop.

So in the shutdown routine, try to shut it down "nicely" via the FS
-stop command first, and then if it isn't dead after 10 seconds kill 
the process with the standard sip_stop escalation procedure.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@15182 ab1d8caa-1f67-47f1-9e81-24633a41865c