sipXconfig: changes for new Java Based VM
Have sipXconfig generate a per mailbox PersonalAttendant.properties
file, that has the same information that is in the older
perl mailbox savemessage.vxml VoiceXML script.

This new file is generated in addition to the existing VoiceXML 
script for now.

Upgrades will require this file to be generated for all mailbox users.
This set of changes does not take this requirement into account.  It 
may be that his happens automatically on --first-run when everything
gets regenerated--or not.

Part of (XECS-1610) Create a new FreeSWITCH based Voicemail 
   system that can replace the current VM

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@15289 ab1d8caa-1f67-47f1-9e81-24633a41865c