XCF-3521 Make dial plan replication and restart more usable
Add a new 'replicateIfNeeded' method to DialPlanActivationManager If
there was a previous request to replicate dial plan, calling this method
will replicate it immediately. If there are no replication requests
waiting (or if we are in eager manager) this is a noop.

SipxProcessManager has a new 'markDialPlanRelatedServicesForRestart'
method.  Services are marked for restart right after replication
request, and context remembers the dial plan needs to be replicated if
any of those services are restarted.

As a result the restart services link now comes up immediatley. When you
restart the service, sipXconfig will replicate the dialplans (if it has
not yet been replicated) before doing so.

Increase the replication wait time to 30 secs (it's better not to push
dial plans and restart to often: we kept it short because people were
waiting for Restart link to show up. Now that it is displayed right away
there is no reason to rush replication.)

Based on patch from: Arjun Nair - thanks!

Merged from: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/4.0@15271

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@15298 ab1d8caa-1f67-47f1-9e81-24633a41865c