XX-5587 make XML model parsing thread safe
We had several instanced of ClassCastExceptions in SettingTypeIdRule
when parsing XML model definitions.

It looks like all the instances of Digester used to share a single
static instance of SettingTypeIdRule (it was declared as a static field
of SettingTypeRule). This is dangerous: if we are parsing more than
single file we will end up sharing types definitions between the files.
That was most probably what caused ClassCastException.

We now construct a fresh instance of SettingTypeIdRule for every parsed
file (just as we construct a fresh instance of Digester and every other
Rule). The 'settign type id' rule is shared by all SettingTypeRules:
it's passed as a constructor parameter.

Patch from: Damian Krzeminski - Thanks!

Merged from: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/4.0@15438

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@15486 ab1d8caa-1f67-47f1-9e81-24633a41865c