XX-5726 Fixed behavior of web certificate CSR generation
The web certificate CSR generation in sipXconfig was only working if the sip domain was the same as the server name.  The code has been changed to explicitly use the server name of the primary location for creating and accessing the web certificate.

Merged from: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/4.0@15524

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@15549 ab1d8caa-1f67-47f1-9e81-24633a41865c