sipXivr: Allow Dial by Name to seach while digits entered.
When dialing by name, a lookup is done on each digit.  If 2 or fewer 
matches are detected, it exits the dialing phase and enters the
selection phase.  If after 7 'letters' there 9 or fewer matches, it
exits the dialing phase and enters the selection phase.

One can always exit the dialing phase at any point by pressing "#".

The search is done on any/all combinations of first/middle/last names,
in any order.  Thus "John Paul Jones" can be found by spelling "john",
"paul" or "jones", or "johnpaul" or "pauljones" or "pauljohn" or "jonespaul"
or ...

The prompt that mentioned "last name first" has been removed, and replaced
with a (hacked) prompt that now says:
  "Please spell the name of the person.  
   Press seven for 'q' and nine for 'z'"

This should probably be replaced with something like:

  "Please spell the first or last name of the person using as many 
   letters as you know.  You'll be prompted when the list of names 
   is narrowed enough.  Press seven for 'q' and nine for 'z'."

Fixes:
  XX-4754 - Allow Dial by Name to seach while digits entered
  
--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@15925 ab1d8caa-1f67-47f1-9e81-24633a41865c