XX-6097 Dimdim integration: URL changes for Host and Join
Modifies join conference link: user is presented with a form that allows
specifying the name.
Administrator can optionally configure external call-in number: it is
displayed by Dimdim UI during the conference.
If DimDim conference is configured the links to start and join the
conference are displayed on a Participants panel.

Patch from: Huijun Yang - thanks!

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@16221 ab1d8caa-1f67-47f1-9e81-24633a41865c