XX-5695: Software Updates screen cannot determine version on SCS or system built from sources
This patch adds a marker for the version line in the sipx-swadmin.py.in script in order for sipxconfig to look for this line in the output file (version.output) returned by sipxsupervisor. This is needed because yum could have additional output lines like the "Loading "fastestmirror" plugin." and others too. This is only needed for the version part because the list of found updates is created already using some markers
-getCurrentVersion will never return null (otherwise checking for current version will be done at every page refresh(XML-RPC call) - this causes for AutoRefresh component to never display the refreshing gif image)
-Check for updates button is hidden when current version cannot be retrieved
I think that there is no point to let the user update the system when we don't know what version to update
-if current version cannot be retrieved - the error message is displayed in red color
ALSO: refresh component is only displayed when currentVersion is valid. Otherwise only an error message is displayed

Based on patch from: Bogdan Brezoi - Thanks!


Merged from: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/4.0@15873

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@16289 ab1d8caa-1f67-47f1-9e81-24633a41865c