sipXivr: Add new Music On Hold functionality
Adds server side support for:
   XX-4752 - Enh: Add a MOH server in our FreeSWITCH server
   XX-4835 - Multiple music-on-hold source URIs
   XX-6285 - Provide option of external music source for MOH

Adds a new action value to the sipXivr "action" URL parameter
   action=moh

Adds a new parameter "moh" with values:
     (empty)  use original park server music (/var/sipxdata/parkserver/music/)
     l        use local_stream://moh (defined in local_stream.conf.xml)
     p        use portaudio_stream:// (defined in portaudio.conf.xml)
     u{user}  use per user music (defined somewhere TBD.  validusers.xml?)

The "l" value uses the default MOH stream defined in FreeSWITCH, which
can be a directory of WAV files that is rotated through.  Our current
FreeSWITCH config defines that directory to be:
   ..../share/www/doc/stdprompts/music/8000

The "p" values uses the portaudio_stream which is the input to the
soundcard on the FreeSWITCH machine.  This allows for external audio
sources that can feed the soundcard.  The default setup should work 
for most audio hardware without additional configuration.

The "u{user}" is a way of specifying "use MOH source specified by {user}".
There is no current definition, and the code currently loops through audio 
files found in the {mailstore}/moh directory for that user.

The mappingrules.xml file can be modified to use this new server by
changing the entry that points to the park server, and use this instead:

    <userMatch>
      <!--Music on Hold-->
      <description>Forward music-on-hold calls to IVR</description>
      <userPattern>~~mh~.</userPattern>
      <permissionMatch>
        <transform>
          <user>IVR</user>
          <host>{fs host}:{fs port}</host>
          <urlparams>action=moh</urlparams>
          <urlparams>moh={vdigits}</urlparams>
        </transform>
      </permissionMatch>
    </userMatch>

Then the user's phone can be configured to use the ~~mh~{moh} pattern,
where moh is one of the values described above.  This allows each phone
to specify a specific moh "user" value, which is transformed by mappingrules
into the correct URL for the MOH service inside sipXivr.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@16337 ab1d8caa-1f67-47f1-9e81-24633a41865c