XX-6234 minor ITSP configuration changes
Moved "register on initialization" field next to account credentials.
It's still in the Advanced section for all ITSP templates (which should
just specify it correctly). It's in the Basic section for 'generic'
ITSP.

Enabled by default Music on Hold for sipXbridge: recent sipXbridge
changes allow for enabling MoH on the phone and in sipXbridge at the
same time.

Patch from: Bogdan Brezoi - thanks!

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@16387 ab1d8caa-1f67-47f1-9e81-24633a41865c