XX-6407: Validation's message is not displayed in UserPortal\My Information
Noticed that AutorefreshForm's validator is not the same instance as the contained page validator.
The error gets saved in autorefresh form validator, but it should have been saved in page's validator instead.
The page is responsible with error display, not the autorefresh component
The problem is that this is not possible to "delegate" the page's validator to autorefresh form,
because @Bean annotated methods does not support automatic binding; instead it creates a new object
instance where it is defined.
I didn't expect to this kind of behaviour, but whenever I tried to perform a "set" call an exception
was thrown by tapestry saying that the operation is not permitted
The only solution was to manually record firstError (wich is displayed by ErrorMsg component) from
autorefresh form's validator to its contained page validator.
EditBridge error.connect message moved from conferencesPanel component to EditBridge.properties
I also searched through pages where autorefreshForm is used, and I noticed that the error is manually recorded
in page's validator; in this case this is a spring validator in action (a bean) that automatically records the error in
the autorefresh component's validator, and I had to get the error from here and send it to the
contained page's validator.

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@16437 ab1d8caa-1f67-47f1-9e81-24633a41865c