XX-6523 generate 'first' and 'last' CSS classes
Standard ":first-child" and ":last-child" pseudo CSS classes are poorly
supported.
sipXconfig will generate 'first' and 'last' class for the <tr> elements
of the table to allow for different styles for the first and last row.

Since there is no easy way of detecting when Tapestry generates the last
row in the table, appending CSS classes is handled by a javascript
automatically attached to Table component.

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@16454 ab1d8caa-1f67-47f1-9e81-24633a41865c