sipXivr: Allow "0" to be used as PIN everywhere
sipXconfig could set "0" as PIN, and a user could log in with it, but
then wouldn't be able to change their PIN as entering "0" would
divert the call to the operator.  If their PIN was something other than 0,
then they couldn't set their PIN to 0, as when they entered that the
call would be diverted to the operator.

This fix makes it consistent:
  "0" as a PIN is allowed, and will NOT divert to operator if entered.

Fixes:
   XX-6555 - Voicemail: handling of "0" as PIN is inconsistent

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@16478 ab1d8caa-1f67-47f1-9e81-24633a41865c