sipXcommons, sipXivr: Have the voice mail notification e-mail use a template instead of hardcoding it
Adds support for three "types" of formats:
  Full:  Same as the old format, HTML and text alterntatives, with links.
  Medium: HTML and text alternatves, without links.
  Brief: Just text.

Templates are defined in EmailFormats.properites, which can be overridden
with a user supplied version, and also internationalized.  Lots of 
variable expansions available to enable custom formats of items such
as the subject, date formatting, etc.

sipXivr will look for {prefix}/etc/sipxpbx/sipxivr/EmailFormats.properties
(including all EmailFormats_{locale}.properties variants) first, and if
not found will use the default formats defined in 
org.sipfoundry.voicemail.EmailFormats.properties (in the sipXivr.jar file).

As there is no sipXconfig support for configuring this yet, the default
format for the main email address is Full, and for the alternate email
address is Brief at the moment.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@16908 ab1d8caa-1f67-47f1-9e81-24633a41865c