sipXconfig: Change MOH config slightly
Changed MOH config from:
    <directory name="moh/8000" path="${moh.audioDirectoryPath}">
to
    <directory name="moh" path="${moh.audioDirectoryPath}">

When FreeSWITCH is told to play the localstream named "moh", it
attempts to match a directory name with the sample rate of the current 
codec for the call, so as to find the best match and not have
to transcode.  When a call comes in it tries "moh/{samplerate}", then
and if it cannot find that directory, it then tries just "moh".

With the previous config, if call came in with G.722 (16 Khz sample
rate), then the first search was for "moh/16000", which it didn't find,
so then it tried "moh", which it also didn't find.  It would end up
using "default" instead.

I also disabled the "shuffle" mode, because with a small number of
files in the directory, it might keep picking the same file (random
chance), and it appeared that MOH wasn't playing correctly.  Rather
than deal with customer calls on that, I just disabled it.

Fixes XX-6964 - Uploaded MOH file is failed to play on Polycom phones
Fixes XX-6984 - Intermittent :Only single MOH is heard on first participant of the conference after uploading several MOH files to the system music directory

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@17184 ab1d8caa-1f67-47f1-9e81-24633a41865c