XX-7386: After adding new CA a way is needed to apply from GUI (i.e. without accessing ssh/console)
-When CA gets changed, the mark services for restart mechanism is used. For the moment only
configuration and bridge services are marked for restart. With regard to google accounts import, technically, only the configuration
service needs restart. SipxBridge service may use CA too and therefore I marked it for restart as well.
If one decides that sipXBridge is not affected by CA change, can ignore to restart it.
-RestartNeededServicesPage is improved to show WaitingPage if Configuration service is marked for restart
and the user chooses to restart it
-Integration tests added to test marked services for restart being given a locations configuration
-Test added to verify when Server restart is occuring

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@17757 ab1d8caa-1f67-47f1-9e81-24633a41865c