XX-4911 - forced a replication after the bridge is deleted when the location is saved.
 - Needed to change a bit the deploy() method from deploy(int) to deploy(Bridge) in the bridge provisioning mechanism
The id of the bridge was used to retrieve the bridge from Hibernate and then the bridge to retrieve the FS service and the location in order to make the replication. We needed to skip the retrieval of the bridge from the DB since the bridge is deleted from DB before the replication being triggered.

 - cleanup - removed an unused method

Patch from Cristi Starasciuc. Thanks!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@17828 ab1d8caa-1f67-47f1-9e81-24633a41865c