XX-7213: HA - Install of new certificate authorities needed on Distributed Servers as well
Changes will do an immediate replicate of the CA to all remote servers and then restart (not reboot) the remote servers in order for the new CA to be recognized
Changes on top of Raymond's patch:
-The replication mechanism in sipXconfig is generic enough to support any file replication
(not only specific configuration or xml file types)
So there is no need to add new replication mechanism
-Added AnyFile.java that provides mechanism for any file replication (applicable to certificate files also)
-Similar to other file replications, I choosed to replicate the certificate on all locations instead of
simply copying the file on primary location and replicate on others.

Based on patch from: Raymond Dans - Thanks!

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@18103 ab1d8caa-1f67-47f1-9e81-24633a41865c