XX-7655 Call filters are not working as expected in CDR historic page
Expand call filtering to include additional call tags which represent internal destinations

Config CDR Search was too restrictive for searching on To field. Updated the SQL Search for this condition to include tests for a number of different route tags that indicate an internal destination.

Please note: That for external calls (i.e. PSTN type calls), the search will still not be supported as a decision was made awhile ago to only support internal.

Patch from Raymond Dans, thanks!

git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@18269 ab1d8caa-1f67-47f1-9e81-24633a41865c